+++ tests/v1/integration/application-charm/src/etcd_requires.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright 2022 Canonical Ltd.
+# Copyright 2026 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 """Etcd requires interface implementation using data interfaces v1."""
